+++ JobMatchBackend/Repositories/ICompanyRepository.cs
@@ -7,6 +7,5 @@ public interface ICompanyRepository
 {
     Task<User?> GetCompanyByIdAsync(Guid companyId);
     Task<int> GetActiveJobsCountAsync(Guid companyId);
-    Task<List<User>> GetAllCompaniesAsync();  
-
+    Task<List<User>> GetAllAsync();
 }
